rpa-client/requirements.txt
pywinauto>=0.6.8
pyautogui>=0.9.53
PyQt5>=5.15.0
opencv-python>=4.5.0
websockets>=10.0
requests>=2.28.0
pywin32>=305; platform_system=="Windows"
APScheduler>=3.10.0
loguru>=0.7.0
Pillow>=9.0.0
numpy>=1.21.0
cryptography>=41.0.0
uuid
